GenId

API Documentation for GenId.

GenId.bit_mask_int

GenId.bit_mask_uint

GenId.bit_mask_uint

GenId.tsid_from_string

GenId.tsid_generate

GenId.tsid_machine_id

GenId.tsid_machine_tail

GenId.tsid_timestamp

GenId.bit_mask_int — Method

bit_mask_int(type, v, from, to)

Applies a 64 bit signed integer mask to a value v, with ones between bits in postitions between from and to and zeroes for other bits.

Counting of bits starts at 0.

v can't be a negative integer.

Examples

julia> bit_mask_int(Int64, typemax(Int64), 0,0)
1
julia> bit_mask_int(Int64, typemax(Int64), 0,1)
3

GenId.bit_mask_uint — Method

bit_mask_uint(type, v, from, to)

Masks a 64 bit unsigned integer v with ones between bits in postitions between from and to and zeroes for other bits.

Counting of bits starts at 0.

Examples

julia> GenId.bit_mask_uint(UInt64, typemax(UInt64), 0,0)
0x0000000000000001
julia> GenId.bit_mask_uint(UInt64, typemax(UInt64), 12, 21)
0x00000000003ff000

GenId.bit_mask_uint — Method

bit_mask_uint(from, to)

Creates a 64 bit unsigned integer mask, with ones between bits in postitions between from and to and zeroes for other bits.

Counting of bits starts at 0.

Examples

julia> bit_mask_uint(0,0)
0x0000000000000000000000000000000000000000000000000000000000000001
julia> bit_mask_uint(63,63)
0x1000000000000000000000000000000000000000000000000000000000000000
julia> bit_mask_uint(12,21)
0x0000000000000000000000000000000000000000001111111111000000000000

GenId.tsid_from_string — Method

tsid_from_string(s::AbstractString)

Creates a new UUID from a textual representation in s.

Examples

julia> tsid_from_string("DJR0RGDG0401")
489485826766409729

GenId.tsid_generate — Method

tsid_generate(def)

Creates a new UUID based on def.

Examples

julia> iddef = TsIdDefinition(
    # Data type used for storage of the ID
    Int64; 
    # Number of bits used for the timestamp section.
    bits_time=41, 
    # Number of bits used for the machine section.
    bits_machine=10, 
    # Number of bits for the tail section.
    # Can be a random number or a local machine/thread specific sequence.
    bits_tail=12, 
    machine_id=1, 
    # Start of the epoch for this UUID scheme.
    # Time before that can't be represented.
    epoch_start_dt=DateTime(2020, 1, 1, 0, 0, 0, 0), 
    # Start of the epoch for this UUID scheme.
    # Time after that can't be represented.
    epoch_end_dt=DateTime(2070, 12, 31, 23, 59, 59, 999)
)
TsIdDefinition(Int64, 41, 10, 12, 1, 4096, 22, 12, Dates.DateTime("2020-01-01T00:00:00"), Dates.DateTime("2070-12-31T23:59:59.999"), 63713520000000, 65322979199999)

julia> tsid_generate(iddef)
489485826766409729

GenId.tsid_machine_id — Method

tsid_machine_id(def::TsIdDefinition, tsid::TT) where {TT<:Integer}

Extracts the machine id from an existing tsid.

Examples

julia> tsid_machine_id(iddef, 489485826766409729)
1

GenId.tsid_machine_tail — Method

tsid_machine_tail(def::TsIdDefinition, tsid::TT) where {TT<:Integer}

Extracts the tail (an increment or a random number) from an existing tsid.

Examples

julia> tsid_machine_tail(iddef, 489485826766409729)
1

GenId.tsid_timestamp — Method

tsid_timestamp(def::TsIdDefinition, tsid::TT) where {TT<:Integer}

Extracts the timestamp from an existing tsid.

Examples

julia> tsid_timestamp(iddef, 489485826766409729)
2023-09-12T17:21:55.308